+++ update-run.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 
 outFilename="out1"
-(cd build; make;) && time ./build/raytracer scene.yaml -o results/$outFilename.png | tee raytracer.log
+(cd build; make;) && time ./build/raytracer scene_exp.yaml -o results/$outFilename.png | tee raytracer.log
 # convert $outFilename.ppm $outFilename.png