+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Forthcoming
 -----------
+- FEAT: Added support for anti aliasing pattern in camera
 
-    
 2018/02/11 - Added scene loading
 --------------------------------
 - Added changelog (http://keepachangelog.com)